+++ src/poetry/installation/chef.py
@@ -1,105 +1,26 @@
 from __future__ import annotations
 
-import os
 import tempfile
 
-from contextlib import redirect_stdout
-from io import StringIO
 from pathlib import Path
 from typing import TYPE_CHECKING
 
 from build import BuildBackendException
-from build import ProjectBuilder
-from build.env import IsolatedEnv as BaseIsolatedEnv
 from poetry.core.utils.helpers import temporary_directory
-from pyproject_hooks import quiet_subprocess_runner  # type: ignore[import-untyped]
 
 from poetry.utils._compat import decode
-from poetry.utils.env import ephemeral_environment
 from poetry.utils.helpers import extractall
+from poetry.utils.isolated_build import IsolatedBuildError
+from poetry.utils.isolated_build import isolated_builder
 
 
 if TYPE_CHECKING:
-    from collections.abc import Collection
-
     from poetry.repositories import RepositoryPool
     from poetry.utils.cache import ArtifactCache
     from poetry.utils.env import Env
 
 
 class ChefError(Exception): ...
-
-
-class ChefBuildError(ChefError): ...
-
-
-class ChefInstallError(ChefError):
-    def __init__(self, requirements: Collection[str], output: str, error: str) -> None:
-        message = "\n\n".join(
-            (
-                f"Failed to install {', '.join(requirements)}.",
-                f"Output:\n{output}",
-                f"Error:\n{error}",
-            )
-        )
-        super().__init__(message)
-        self._requirements = requirements
-
-    @property
-    def requirements(self) -> Collection[str]:
-        return self._requirements
-
-
-class IsolatedEnv(BaseIsolatedEnv):
-    def __init__(self, env: Env, pool: RepositoryPool) -> None:
-        self._env = env
-        self._pool = pool
-
-    @property
-    def python_executable(self) -> str:
-        return str(self._env.python)
-
-    def make_extra_environ(self) -> dict[str, str]:
-        path = os.environ.get("PATH")
-        scripts_dir = str(self._env._bin_dir)
-        return {
-            "PATH": (
-                os.pathsep.join([scripts_dir, path])
-                if path is not None
-                else scripts_dir
-            )
-        }
-
-    def install(self, requirements: Collection[str]) -> None:
-        from cleo.io.buffered_io import BufferedIO
-        from poetry.core.packages.dependency import Dependency
-        from poetry.core.packages.project_package import ProjectPackage
-
-        from poetry.config.config import Config
-        from poetry.installation.installer import Installer
-        from poetry.packages.locker import Locker
-        from poetry.repositories.installed_repository import InstalledRepository
-
-        # We build Poetry dependencies from the requirements
-        package = ProjectPackage("__root__", "0.0.0")
-        package.python_versions = ".".join(str(v) for v in self._env.version_info[:3])
-        for requirement in requirements:
-            dependency = Dependency.create_from_pep_508(requirement)
-            package.add_dependency(dependency)
-
-        io = BufferedIO()
-        installer = Installer(
-            io,
-            self._env,
-            package,
-            Locker(self._env.path.joinpath("poetry.lock"), {}),
-            self._pool,
-            Config.create(),
-            InstalledRepository.load(self._env),
-        )
-        installer.update(True)
-        if installer.run() != 0:
-            raise ChefInstallError(requirements, io.fetch_output(), io.fetch_error())
 
 
 class Chef:
@@ -127,46 +48,36 @@ class Chef:
     ) -> Path:
         from subprocess import CalledProcessError
 
-        with ephemeral_environment(
-            self._env.python,
-            flags={"no-pip": True, "no-setuptools": True, "no-wheel": True},
-        ) as venv:
-            env = IsolatedEnv(venv, self._pool)
-            builder = ProjectBuilder.from_isolated_env(
-                env, directory, runner=quiet_subprocess_runner
-            )
-            env.install(builder.build_system_requires)
+        distribution = "wheel" if not editable else "editable"
+        error: Exception | None = None
 
-            stdout = StringIO()
-            error: Exception | None = None
-            try:
-                with redirect_stdout(stdout):
-                    dist_format = "wheel" if not editable else "editable"
-                    env.install(
-                        builder.build_system_requires
-                        | builder.get_requires_for_build(dist_format)
+        try:
+            with isolated_builder(
+                source=directory,
+                distribution=distribution,
+                python_executable=self._env.python,
+                pool=self._pool,
+            ) as builder:
+                return Path(
+                    builder.build(
+                        distribution,
+                        destination.as_posix(),
                     )
-                    path = Path(
-                        builder.build(
-                            dist_format,
-                            destination.as_posix(),
-                        )
-                    )
-            except BuildBackendException as e:
-                message_parts = [str(e)]
-                if isinstance(e.exception, CalledProcessError):
-                    text = e.exception.stderr or e.exception.stdout
-                    if text is not None:
-                        message_parts.append(decode(text))
-                else:
-                    message_parts.append(str(e.exception))
+                )
+        except BuildBackendException as e:
+            message_parts = [str(e)]
 
-                error = ChefBuildError("\n\n".join(message_parts))
+            if isinstance(e.exception, CalledProcessError):
+                text = e.exception.stderr or e.exception.stdout
+                if text is not None:
+                    message_parts.append(decode(text))
+            else:
+                message_parts.append(str(e.exception))
 
-            if error is not None:
-                raise error from None
+            error = IsolatedBuildError("\n\n".join(message_parts))
 
-            return path
+        if error is not None:
+            raise error from None
 
     def _prepare_sdist(self, archive: Path, destination: Path | None = None) -> Path:
         from poetry.core.packages.utils.link import Link
