+++ tests/installation/test_chef.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import os
 import shutil
-import sys
 import tempfile
 
 from pathlib import Path
@@ -16,18 +15,11 @@ from poetry.core.packages.utils.link import Link
 
 from poetry.factory import Factory
 from poetry.installation.chef import Chef
-from poetry.installation.chef import ChefInstallError
-from poetry.installation.chef import IsolatedEnv
-from poetry.puzzle.exceptions import SolverProblemError
-from poetry.puzzle.provider import IncompatibleConstraintsError
 from poetry.repositories import RepositoryPool
 from poetry.utils.env import EnvManager
-from poetry.utils.env import ephemeral_environment
 
 
 if TYPE_CHECKING:
-    from collections.abc import Collection
-
     from pytest_mock import MockerFixture
 
     from poetry.repositories.pypi_repository import PyPiRepository
@@ -48,41 +40,6 @@ def pool(pypi_repository: PyPiRepository) -> RepositoryPool:
 @pytest.fixture(autouse=True)
 def setup(mocker: MockerFixture, pool: RepositoryPool) -> None:
     mocker.patch.object(Factory, "create_pool", return_value=pool)
-
-
-def test_isolated_env_install_success(pool: RepositoryPool) -> None:
-    with ephemeral_environment(Path(sys.executable)) as venv:
-        env = IsolatedEnv(venv, pool)
-        assert "poetry-core" not in venv.run("pip", "freeze")
-        env.install({"poetry-core"})
-        assert "poetry-core" in venv.run("pip", "freeze")
-
-
-@pytest.mark.parametrize(
-    ("requirements", "exception"),
-    [
-        ({"poetry-core==1.5.0", "poetry-core==1.6.0"}, IncompatibleConstraintsError),
-        ({"black==19.10b0", "attrs==17.4.0"}, SolverProblemError),
-    ],
-)
-def test_isolated_env_install_error(
-    requirements: Collection[str], exception: type[Exception], pool: RepositoryPool
-) -> None:
-    with ephemeral_environment(Path(sys.executable)) as venv:
-        env = IsolatedEnv(venv, pool)
-        with pytest.raises(exception):
-            env.install(requirements)
-
-
-def test_isolated_env_install_failure(
-    pool: RepositoryPool, mocker: MockerFixture
-) -> None:
-    mocker.patch("poetry.installation.installer.Installer.run", return_value=1)
-    with ephemeral_environment(Path(sys.executable)) as venv:
-        env = IsolatedEnv(venv, pool)
-        with pytest.raises(ChefInstallError) as e:
-            env.install({"a", "b>1"})
-        assert e.value.requirements == {"a", "b>1"}
 
 
 def test_prepare_sdist(
