+++ tests/installation/test_executor.py
@@ -1263,7 +1263,7 @@ Package operations: 1 install, 0 updates, 0 removals
 
   - Installing {package_name} ({package_version} {package_url})
 
-  ChefBuildError
+  IsolatedBuildError
 
   hide the original error
   \
@@ -1409,7 +1409,7 @@ Package operations: 1 install, 0 updates, 0 removals
 
   - Installing {package_name} ({package_version} {package_url})
 
-  ChefInstallError
+  IsolatedBuildInstallError
 
   Failed to install poetry-core>=1.1.0a7.
   \
@@ -1426,6 +1426,7 @@ Package operations: 1 install, 0 updates, 0 removals
 
     mocker.stopall()  # to get real output
     output = io.fetch_output().strip()
+    print(output)
     assert output.startswith(expected_start)
     assert output.endswith(expected_end)
